Throw NotFoundException (404) when an unknown tenant is referenced.

+++ doradus-server/src/main/java/com/dell/doradus/service/tenant/TenantService.java
@@ -35,6 +35,7 @@ import com.dell.doradus.service.db.DColumn;
 import com.dell.doradus.service.db.DuplicateException;
 import com.dell.doradus.service.db.Tenant;
 import com.dell.doradus.service.db.UnauthorizedException;
+import com.dell.doradus.service.rest.NotFoundException;
 import com.dell.doradus.service.rest.RESTCommand;
 import com.dell.doradus.service.rest.RESTService;
 import com.dell.doradus.service.schema.SchemaService;
@@ -213,7 +214,9 @@ public class TenantService extends Service {
         Utils.require(ServerConfig.getInstance().multitenant_mode,
                       "This command is not valid in single-tenant mode");
         TenantDefinition tenantDef = getTenantDefFromCache(tenantName);
-        Utils.require(tenantDef != null, "Tenant '%s' does not exist", tenantName);
+        if (tenantDef == null) {
+            return; // allow idempotent deletes
+        }
         Tenant tenant = new Tenant(tenantName);
         DBService.instance().dropTenant(tenant);
         deleteTenantFromCache(tenantName);
@@ -281,17 +284,23 @@ public class TenantService extends Service {
      * represented by the given userID and password. A {@link UnauthorizedException} is
      * thrown if the given credentials are invalid for the given tenant/command or if the
      * user has insufficient rights for the command being accessed.
+     * <p>
+     * If the given tenant does not exist, an {@link NotFoundException} is thrown.
      * 
      * @param tenant        {@link Tenant} being accessed.
      * @param userid        User ID of caller. May be null.
      * @param password      Password of caller. May be null.
      * @param cmd           {@link RESTCommand} being invoked.
+     * @throws              NotFoundException if the given tenant does not exist.
      * @throws              UnauthorizedException if the given credentials are invalid or
      *                      have insufficient rights for the given tenant and command.
      */
     public void validateTenantAuthorization(Tenant tenant, String userid, String password, RESTCommand cmd)
-            throws UnauthorizedException {
+            throws UnauthorizedException, NotFoundException {
         if (ServerConfig.getInstance().multitenant_mode) {
+            if (!tenantExists(tenant)) {
+                throw new NotFoundException("Unknown tenant: " + tenant);
+            }
             if (cmd.isPrivileged()) {
                 if (!isValidSystemCredentials(userid, password)) {
                     throw new UnauthorizedException("Unrecognized system user id/password");
